+++ sdk/servicebus/service-bus/src/serviceBusRuleManager.ts
@@ -15,7 +15,7 @@ import { getUniqueName } from "./util/utils";
 import { throwErrorIfConnectionClosed } from "./util/errors";
 import { SqlRuleFilter } from "./serializers/ruleResourceSerializer";
 import { tracingClient } from "./diagnostics/tracing";
-import { getPagedAsyncIterator, PagedAsyncIterableIterator, PagedResult, PageSettings } from "@azure/core-paging";
+import { getPagedAsyncIterator, PagedAsyncIterableIterator, PagedResult } from "@azure/core-paging";
 import { OperationOptions } from "@azure/core-client";
 import { ListRequestOptions } from "./serviceBusAtomManagementClient";
 
@@ -65,10 +65,6 @@ export interface ServiceBusRuleManager {
    * @returns An asyncIterableIterator that supports paging.
    */
   listRules(
-    // eslint-disable-next-line @azure/azure-sdk/ts-naming-options
-    options?: OperationOptions
-  ): PagedAsyncIterableIterator<RuleProperties>;
-  listRules2(
     // eslint-disable-next-line @azure/azure-sdk/ts-naming-options
     options?: OperationOptions
   ): PagedAsyncIterableIterator<RuleProperties>;
@@ -204,57 +200,6 @@ export class ServiceBusRuleManagerImpl implements ServiceBusRuleManager {
     );
   }
 
-  public listRules2(
-    options?: OperationOptions
-  ): PagedAsyncIterableIterator<RuleProperties, RuleProperties[], { maxPageSize?: number }> {
-    const that = this;
-    const pagedResult: PagedResult<RuleProperties[], { maxPageSize?: number }, number> = {
-      firstPageLink: 0,
-      async getPage(pageLink, maxPageSize) {
-        const top = maxPageSize ?? 100;
-        const rules = await that.getRules({
-          skip: pageLink,
-          maxCount: top,
-          ...options,
-        });
-        return {
-          page: rules,
-          nextPageLink: rules.length > 0 ? pageLink + rules.length: undefined
-        }
-      }
-    };
-
-    return getPagedAsyncIterator(pagedResult)
-  }
-
-  private async *listRulesPage(
-    marker?: string,
-    options: OperationOptions & Pick<PageSettings, "maxPageSize"> = {}
-  ): AsyncIterableIterator<RuleProperties[]> {
-    do {
-      const rules = await this.getRules({
-        skip: Number(marker),
-        maxCount: options.maxPageSize ?? 100,
-        ...options,
-      });
-      if (rules.length > 0) {
-        yield rules;
-        marker = String(Number(marker ?? 0) + rules.length);
-      } else {
-        break;
-      }
-    } while (marker);
-  }
-
-  private async *listRulesAll(
-    options: OperationOptions = {}
-  ): AsyncIterableIterator<RuleProperties> {
-    let marker: string | undefined;
-    for await (const segment of this.listRulesPage(marker, options)) {
-      yield* segment;
-    }
-  }
-
   /**
    * Returns an async iterable iterator to list all the rules
    * under the specified subscription.
@@ -268,27 +213,25 @@ export class ServiceBusRuleManagerImpl implements ServiceBusRuleManager {
     options?: OperationOptions
   ): PagedAsyncIterableIterator<RuleProperties, RuleProperties[], { maxPageSize?: number }> {
     logger.verbose(`Performing operation - listRules() with options: %j`, options);
-    const iter = this.listRulesAll(options);
-    return {
-      /**
-       */
-      next() {
-        return iter.next();
-      },
-      /**
-       */
-      [Symbol.asyncIterator]() {
-        return this;
-      },
-      /**
-       */
-      byPage: (settings: { maxPageSize?: number } = {}) => {
-        return this.listRulesPage(undefined, {
-          maxPageSize: settings.maxPageSize,
+    const pagedResult: PagedResult<RuleProperties[], { maxPageSize?: number }, number> = {
+      firstPageLink: 0,
+      getPage: async (pageLink, maxPageSize) => {
+        const top = maxPageSize ?? 100;
+        const rules = await this.getRules({
+          skip: pageLink,
+          maxCount: top,
           ...options,
         });
+        return rules.length
+          ? {
+              page: rules,
+              nextPageLink: rules.length > 0 ? pageLink + rules.length : undefined,
+            }
+          : undefined;
       },
     };
+
+    return getPagedAsyncIterator(pagedResult);
   }
 
   /**