+++ sdk/devcenter/arm-devcenter/samples/v1-beta/javascript/galleriesGetSample.js
@@ -15,7 +15,7 @@ const { DefaultAzureCredential } = require("@azure/identity");
  * This sample demonstrates how to Gets a gallery
  *
  * @summary Gets a gallery
- * x-ms-original-file: specification/devcenter/resource-manager/Microsoft.DevCenter/preview/2022-08-01-preview/examples/Galleries_Get.json
+ * x-ms-original-file: specification/devcenter/resource-manager/Microsoft.DevCenter/preview/2022-09-01-preview/examples/Galleries_Get.json
  */
 async function galleriesGet() {
   const subscriptionId = "{subscriptionId}";